backend/requirements.txt
huggingface-hub==0.29.3
python-dotenv==1.1.0